Add RouteCalculator
The RouteCalculator class provides an API to calculate public
transportation routes based on a start position and an end position. It
is a singleton that is initialized at application start, and proptly
builds a directed graph with transportation lines' information. Then, it
provides a method to run an A* pathfinding algorithm in the graph and
find the optimal route.